Se agrega un hook para separar logica
